refactor(giei): cleanup and refactor in giei module
Little refactor and cleanup of old debug comments

+++ src/cores/core_0/feature/giei/giei_components/regen_alg/regen_alg.h
@@ -4,11 +4,11 @@
 #include <stdint.h>
 
 struct RegenAlgInput {
-    const int16_t front_left_velocity;
-    const int16_t front_right_velocity;
-    const int16_t rear_left_velocity;
-    const int16_t rear_right_velocity;
-    const uint16_t battery_pack_tension;
+    const float front_left_velocity;
+    const float front_right_velocity;
+    const float rear_left_velocity;
+    const float rear_right_velocity;
+    const float battery_pack_tension;
 };
 
 void regen_alg_init(void);
